Completed NetMF small. TODO: Refactor and profile

+++ utils/graph.h
@@ -1,13 +1,22 @@
 #pragma once
+#include<string.h>
+#include<fstream>
+#include<iostream>
+#include <bits/stdc++.h> 
+#include<vector>
+#include "graph.h"
+#include<stdio.h>
+#include<cstring>
+#include <iostream>
 class Graph
 {
 	public:
-		float *adj, *degree, volume;
+		double *adj, *degree, volume;
 		int size;
 		float num_edges;
 		bool directed;
 		Graph(int);
-		void add_edge(int, int, float);
+		void add_edge(int, int, double);
 		void print_degree();
 		void print_graph();
 		void info();
